pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gym-kmanip"
version = "0.0.1"
description = "K-Scale Manipulation Gymnasium Environment"
authors = [
    { name = "Hugo Ponte", email = "hugo@kscale.dev" }
]
license = { text = "MIT" }
readme = "README.md"
dependencies = [
    "mujoco>=2.3.7",
    "gymnasium>=0.29.1",
    "dm-control>=1.0.14",
    "h5py>=3.11.0",
    "rerun-sdk>=0.16.0",
    "imageio[ffmpeg]>=2.34.0",
    "pytest>=8.1.0",
    "pyzmq>=26.0.3",
]

[project.urls]
Homepage = "https://github.com/kscale/gym-kmanip"
